requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
celery>=5.3.0
redis>=4.5.0
pillow>=10.0.0
python-dateutil>=2.8.0
django-extensions>=3.2.0

